+++ src/main/java/me/drison64/mypin/managers/ActionsManager.java
@@ -25,8 +25,8 @@
 package me.drison64.mypin.managers;
 
 import me.drison64.mypin.Main;
-import me.drison64.mypin.objects.Action.Action;
-import me.drison64.mypin.objects.Action.ActionType;
+import me.drison64.mypin.objects.action.Action;
+import me.drison64.mypin.objects.action.ActionType;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
