+++ src/main/java/me/drison64/mypin/objects/action/ActionType.java
@@ -22,43 +22,28 @@
  * SOFTWARE.
  */
 
-package me.drison64.mypin.objects.Action;
+package me.drison64.mypin.objects.action;
 
-import me.drison64.mypin.Main;
-import me.drison64.mypin.utils.StringStitcherUtils;
-import org.bukkit.Bukkit;
-import org.bukkit.ChatColor;
-import org.bukkit.block.Block;
-import org.bukkit.entity.Player;
+public enum ActionType {
 
-import java.util.List;
+    INTERACT("interact", Action_Interact.class),
+    WAIT("wait", Action_Wait.class),
+    CONSOLE_COMMAND("console_command", Action_ConsoleCommand.class),
+    PLAYER_COMMAND("player_command", Action_PlayerCommand.class);
 
-public class Action_ConsoleCommand extends Action {
+    private String fancyname;
+    private Class<? extends Action> clazz;
 
-    private Main main;
-    private String[] splitted;
-    private String command;
-    private int delay = 1;
-
-    public Action_ConsoleCommand(Main main, ActionType type) {
-        super(main, type);
-        this.main = main;
+    public String getFancyname() {
+        return fancyname;
     }
 
-    @Override
-    public void run(List<String> data, Integer line, Block block, Player player) {
-
-        this.splitted = data.get(line - 1).split(" ");
-
-        if (splitted.length < 2) {
-            Bukkit.getConsoleSender().sendMessage(ChatColor.DARK_RED + "Error occured at line: " + (line + 1) + ", value is empty.");
-        }
-
-        command = StringStitcherUtils.stitch(splitted, 1);
-
-        Bukkit.dispatchCommand(Bukkit.getConsoleSender(), command);
-        runNext(data, line, block, player);
-
+    public Class<? extends Action> getClazz() {
+        return clazz;
     }
 
+    ActionType(String fancyname, Class<? extends Action> clazz) {
+        this.fancyname = fancyname;
+        this.clazz = clazz;
+    }
 }
